requirements-test.txt
# Test dependencies for Foundation Security Services Setup
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
moto[all]>=4.0.0
boto3>=1.26.0
python-dotenv>=1.0.0
botocore>=1.29.0